// src/schemas/market.ts
import { z } from "zod";

// API: https://api.coingecko.com/api/v3/global
// JSON to Zod: https://rsinohara.github.io/json-to-zod-react

export const Market = z.object({
    data: z.object({
      active_cryptocurrencies: z.number(),
      upcoming_icos: z.number(),
      ongoing_icos: z.number(),
      ended_icos: z.number(),
      markets: z.number(),
      total_market_cap: z.object({
        btc: z.number(),
        eth: z.number(),
        ltc: z.number(),
        bch: z.number(),
        bnb: z.number(),
        eos: z.number(),
        xrp: z.number(),
        xlm: z.number(),
        link: z.number(),
        dot: z.number(),
        yfi: z.number(),
        usd: z.number(),
        aed: z.number(),
        ars: z.number(),
        aud: z.number(),
        bdt: z.number(),
        bhd: z.number(),
        bmd: z.number(),
        brl: z.number(),
        cad: z.number(),
        chf: z.number(),
        clp: z.number(),
        cny: z.number(),
        czk: z.number(),
        dkk: z.number(),
        eur: z.number(),
        gbp: z.number(),
        hkd: z.number(),
        huf: z.number(),
        idr: z.number(),
        ils: z.number(),
        inr: z.number(),
        jpy: z.number(),
        krw: z.number(),
        kwd: z.number(),
        lkr: z.number(),
        mmk: z.number(),
        mxn: z.number(),
        myr: z.number(),
        ngn: z.number(),
        nok: z.number(),
        nzd: z.number(),
        php: z.number(),
        pkr: z.number(),
        pln: z.number(),
        rub: z.number(),
        sar: z.number(),
        sek: z.number(),
        sgd: z.number(),
        thb: z.number(),
        try: z.number(),
        twd: z.number(),
        uah: z.number(),
        vef: z.number(),
        vnd: z.number(),
        zar: z.number(),
        xdr: z.number(),
        xag: z.number(),
        xau: z.number(),
        bits: z.number(),
        sats: z.number(),
      }),
      total_volume: z.object({
        btc: z.number(),
        eth: z.number(),
        ltc: z.number(),
        bch: z.number(),
        bnb: z.number(),
        eos: z.number(),
        xrp: z.number(),
        xlm: z.number(),
        link: z.number(),
        dot: z.number(),
        yfi: z.number(),
        usd: z.number(),
        aed: z.number(),
        ars: z.number(),
        aud: z.number(),
        bdt: z.number(),
        bhd: z.number(),
        bmd: z.number(),
        brl: z.number(),
        cad: z.number(),
        chf: z.number(),
        clp: z.number(),
        cny: z.number(),
        czk: z.number(),
        dkk: z.number(),
        eur: z.number(),
        gbp: z.number(),
        hkd: z.number(),
        huf: z.number(),
        idr: z.number(),
        ils: z.number(),
        inr: z.number(),
        jpy: z.number(),
        krw: z.number(),
        kwd: z.number(),
        lkr: z.number(),
        mmk: z.number(),
        mxn: z.number(),
        myr: z.number(),
        ngn: z.number(),
        nok: z.number(),
        nzd: z.number(),
        php: z.number(),
        pkr: z.number(),
        pln: z.number(),
        rub: z.number(),
        sar: z.number(),
        sek: z.number(),
        sgd: z.number(),
        thb: z.number(),
        try: z.number(),
        twd: z.number(),
        uah: z.number(),
        vef: z.number(),
        vnd: z.number(),
        zar: z.number(),
        xdr: z.number(),
        xag: z.number(),
        xau: z.number(),
        bits: z.number(),
        sats: z.number(),
      }),
      market_cap_percentage: z.object({
        btc: z.number(),
        eth: z.number(),
        usdt: z.number(),
        bnb: z.number(),
        usdc: z.number(),
        xrp: z.number(),
        ada: z.number(),
        steth: z.number(),
        doge: z.number(),
        matic: z.number(),
      }),
      market_cap_change_percentage_24h_usd: z.number(),
      updated_at: z.number(),
    }),
  });